# laravel/task_Agiletech/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/branch": {
            "get": {
                "tags": [
                    "branch"
                ],
                "summary": "Get a list of branchs",
                "operationId": "7a60d8d8fd907b42bfb98cf328e9b958",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/branch/": {
            "post": {
                "summary": "Store a newl branch created resource in storage.",
                "operationId": "8f741dbac90ec38d9a6120088586d2bc",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "branch's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Branch's brand_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "description": "Branch's district_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The branch was successfully created"
                    }
                }
            }
        },
        "/api/branch/id": {
            "get": {
                "tags": [
                    "branch"
                ],
                "summary": "Display the details of branch.",
                "operationId": "87048c3dd53282cf795f4ed0940c7873",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update the  branch in storage.",
                "operationId": "4402c48049025b63beae26c98b5eef84",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Branch's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Branch's brand_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "description": "Branch's district_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img",
                        "in": "query",
                        "description": "Branch's imgs",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The branch was successfully updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "branch"
                ],
                "summary": "Remove the branch from storage.",
                "operationId": "8418fbd188dd44a62582cc494440f857",
                "responses": {
                    "200": {
                        "description": "The branch was successfully deleted"
                    }
                }
            }
        },
        "/api/brand": {
            "get": {
                "tags": [
                    "brand"
                ],
                "summary": "Get a list of brands",
                "operationId": "25c1b1e2bbeba7bfb9df6748b277a205",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/brand/": {
            "post": {
                "summary": "Store a newl brand created resource in storage.",
                "operationId": "9924f278a07cd89508fd63b9caa4148c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "brand's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img",
                        "in": "query",
                        "description": "brand's img",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The brand was successfully created"
                    }
                }
            }
        },
        "/api/brand/id": {
            "get": {
                "tags": [
                    "brand"
                ],
                "summary": "Display the details of brand.",
                "operationId": "a7d753156fcb04c74f1e476a54db37bd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update the  brand in storage.",
                "operationId": "69ec346a10879ad41a8768043de0b0d4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "brand's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "brand's brand_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img",
                        "in": "query",
                        "description": "brand's imgs",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The brand was successfully updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "brand"
                ],
                "summary": "Remove the brand from storage.",
                "operationId": "45118f6b182a1c8cbc812635ab6a6139",
                "responses": {
                    "200": {
                        "description": "The brand was successfully deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}